requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# ML and AutoML
pycaret>=3.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0  # Optional: Advanced gradient boosting
lightgbm>=4.0.0  # Optional: LightGBM for better accuracy
imbalanced-learn>=0.11.0  # Optional: For handling imbalanced datasets (SMOTE)

# Vector search
faiss-cpu>=1.7.4

# LLM integration
transformers>=4.30.0
huggingface-hub>=0.16.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0  # Gemini API

# Dataset sources
kaggle>=1.5.12
datasets>=2.14.0

# GitHub operations
PyGithub>=1.59.0
gitpython>=3.1.31

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

